:art: sqllitedb housekeeping
